Add Vale to the Lint (Docs) workflow
Install and run Vale in the Lint (Docs) workflow in order to ensure that
the documentation has a consistent style. This approach downloads a
release from the Vale GitHub repository, extracts a precompiled binary,
and runs it on all docs files. It assumes that all Vale warnings and
errors are already resolved, so that any new ones would have been
introduced by the target pull request.

This is an alternative to using reviewdog, which was behind a security
incident (see #53162 for context). While reviewdog can inspect the diff
introduced by a PR and limit its checks to changed lines, this change
does not require reliance on an insecure dependency.